# [WIP] Simplifying stage 2 - more flexible code