--- a/include/simdjson/jsonparser.h
+++ b/include/simdjson/jsonparser.h
@@ -11,6 +11,8 @@
 #include <string>
 
 namespace simdjson {
+
+// (This is not meant for end users, see json_parse.)
 // json_parse_implementation is the generic function, it is specialized for
 // various architectures, e.g., as
 // json_parse_implementation<Architecture::HASWELL> or
@@ -45,148 +47,7 @@ int json_parse_implementation(const uint8_t *buf, size_t len, ParsedJson &pj,
   }
   return res;
 }
-namespace {
 
-typedef struct {
-  size_t last_index;
-  size_t last_buf_index;
-  bool has_error;
-} last_index_t;
-
-last_index_t find_last_open_brace(const uint8_t *buf, size_t len,
-                                  ParsedJson &pj) {
-  last_index_t answer;
-  if (pj.n_structural_indexes == 0) {
-    answer.has_error = true;
-    return answer;
-  }
-  uint32_t largest = pj.n_structural_indexes - 1;
-  if (pj.structural_indexes[largest] == len) {
-    if (largest == 0) {
-      answer.has_error = true;
-      return answer;
-    }
-    largest--;
-  }
-  size_t i = largest;
-  for (; i >= 1; i--) {
-    char c = buf[pj.structural_indexes[i]];
-    if ((c == '{') || (c == '[')) {
-      break;
-    }
-  }
-  if (i == 0) {
-    answer.has_error = true;
-    return answer;
-  }
-  answer.has_error = false;
-  answer.last_index = i;
-  answer.last_buf_index = pj.structural_indexes[i];
-  return answer;
-}
-}
-
-// interleaved_json_parse_implementation is the generic function, it is
-// specialized for
-// various architectures, e.g., as
-// interleaved_json_parse_implementation<Architecture::HASWELL> or
-// interleaved_json_parse_implementation<Architecture::ARM64>
-template <Architecture T>
-int interleaved_json_parse_implementation(const uint8_t *buf, size_t len,
-                                          ParsedJson &pj, size_t window,
-                                          bool realloc_if_needed = true) {
-  if (pj.byte_capacity < len) {
-    pj.error_code = simdjson::CAPACITY;
-    return pj.error_code;
-  }
-  if (window <= 0) {
-    pj.error_code = simdjson::WINDOW_ERROR;
-    return pj.error_code;
-  }
-  bool reallocated = false;
-  if (realloc_if_needed) {
-    const uint8_t *tmp_buf = buf;
-    buf = (uint8_t *)allocate_padded_buffer(len);
-    if (buf == NULL)
-      return simdjson::MEMALLOC;
-    memcpy((void *)buf, tmp_buf, len);
-    reallocated = true;
-  } // if(realloc_if_needed) {
-
-  size_t idx = 0;
-  size_t actual_window = window > len ? len : window;
-  bool last_window = false;
-  if (window >= len)
-    last_window = true;
-  int stage1_is_ok =
-      find_structural_bits<T>(buf, actual_window, pj, !last_window);
-  if (stage1_is_ok != simdjson::SUCCESS) {
-    pj.error_code = stage1_is_ok;
-    if (reallocated) {
-      aligned_free((void *)buf);
-    }
-    return pj.error_code;
-  }
-  int stage2_is_ok = unified_machine_init(buf, actual_window, pj);
-  if (stage2_is_ok != simdjson::SUCCESS_AND_HAS_MORE) {
-    pj.error_code = stage2_is_ok;
-    if (reallocated) {
-      aligned_free((void *)buf);
-    }
-    return pj.error_code;
-  }
-  while (idx + window < len) {
-    actual_window = window;
-    stage1_is_ok = find_structural_bits<T>(buf + idx, actual_window, pj, true);
-    if (stage1_is_ok != simdjson::SUCCESS) {
-      pj.error_code = stage1_is_ok;
-      if (reallocated) {
-        aligned_free((void *)buf);
-      }
-      return pj.error_code;
-    }
-    last_index_t last_open_brace_index =
-        find_last_open_brace(buf + idx, actual_window, pj);
-    if (last_open_brace_index.has_error) {
-      pj.error_code = simdjson::WINDOW_ERROR;
-      if (reallocated) {
-        aligned_free((void *)buf);
-      }
-      return pj.error_code;
-    }
-    actual_window = last_open_brace_index.last_buf_index;
-    stage2_is_ok = unified_machine_continue<T>(
-        buf + idx, actual_window, pj, last_open_brace_index.last_index);
-    if (stage2_is_ok != simdjson::SUCCESS_AND_HAS_MORE) {
-      pj.error_code = stage2_is_ok;
-      if (reallocated) {
-        aligned_free((void *)buf);
-      }
-      return stage2_is_ok;
-    }
-    idx += actual_window;
-  }
-  if (idx < len) {
-    actual_window = len - idx;
-    stage1_is_ok = find_structural_bits<T>(buf + idx, actual_window, pj, false);
-
-    if (stage1_is_ok != simdjson::SUCCESS) {
-      pj.error_code = stage1_is_ok;
-      if (reallocated) {
-        aligned_free((void *)buf);
-      }
-      return pj.error_code;
-    }
-    stage2_is_ok =
-        unified_machine_finish<T>(buf + idx, actual_window, pj);
-    idx += actual_window;
-  }
-  pj.error_code = stage2_is_ok;
-  if (reallocated) {
-    aligned_free((void *)buf);
-  }
-  return pj.error_code;
-}
 
 
 // Parse a document found in buf.
@@ -215,6 +76,7 @@ int interleaved_json_parse_implementation(const uint8_t *buf, size_t len,
 
 int json_parse(const uint8_t *buf, size_t len, ParsedJson &pj,
                bool realloc_if_needed = true);
+
 
 // Parse a document found in buf.
 //
@@ -362,5 +224,171 @@ inline ParsedJson build_parsed_json(const padded_string &s) {
   return build_parsed_json(s.data(), s.length(), false);
 }
 
+
+///////////////////////////
+// Interleave
+///////////////////////////
+
+namespace {
+
+typedef struct {
+  size_t last_index;
+  size_t last_buf_index;
+  bool has_error;
+} last_index_t;
+
+last_index_t find_last_open_brace(const uint8_t *buf, size_t len,
+                                  ParsedJson &pj) {
+  last_index_t answer;
+  if (pj.n_structural_indexes == 0) {
+    answer.has_error = true;
+    return answer;
+  }
+  uint32_t largest = pj.n_structural_indexes - 1;
+  if (pj.structural_indexes[largest] == len) {
+    if (largest == 0) {
+      answer.has_error = true;
+      return answer;
+    }
+    largest--;
+  }
+  size_t i = largest;
+  for (; i >= 1; i--) {
+    char c = buf[pj.structural_indexes[i]];
+    if ((c == '{') || (c == '[')) {
+      break;
+    }
+  }
+  if (i == 0) {
+    answer.has_error = true;
+    return answer;
+  }
+  answer.has_error = false;
+  answer.last_index = i;
+  answer.last_buf_index = pj.structural_indexes[i];
+  return answer;
+}
+}
+
+// interleaved_json_parse_implementation is the generic function, it is
+// specialized for
+// various architectures, e.g., as
+// interleaved_json_parse_implementation<Architecture::HASWELL> or
+// interleaved_json_parse_implementation<Architecture::ARM64>
+// (This is not meant for end users, see interleaved_json_parse.)
+template <Architecture T>
+int interleaved_json_parse_implementation(const uint8_t *buf, size_t len,
+                                          ParsedJson &pj, size_t window,
+                                          bool realloc_if_needed = true) {
+  if (pj.byte_capacity < len) {
+    pj.error_code = simdjson::CAPACITY;
+    return pj.error_code;
+  }
+  if (window <= 0) {
+    pj.error_code = simdjson::WINDOW_ERROR;
+    return pj.error_code;
+  }
+  bool reallocated = false;
+  if (realloc_if_needed) {
+    const uint8_t *tmp_buf = buf;
+    buf = (uint8_t *)allocate_padded_buffer(len);
+    if (buf == NULL)
+      return simdjson::MEMALLOC;
+    memcpy((void *)buf, tmp_buf, len);
+    reallocated = true;
+  } // if(realloc_if_needed) {
+
+  size_t idx = 0;
+  size_t actual_window = window > len ? len : window;
+  bool last_window = false;
+  if (window >= len)
+    last_window = true;
+  int stage1_is_ok =
+      find_structural_bits<T>(buf, actual_window, pj, !last_window);
+  if (stage1_is_ok != simdjson::SUCCESS) {
+    pj.error_code = stage1_is_ok;
+    if (reallocated) {
+      aligned_free((void *)buf);
+    }
+    return pj.error_code;
+  }
+  int stage2_is_ok = unified_machine_init(buf, actual_window, pj);
+  if (stage2_is_ok != simdjson::SUCCESS_AND_HAS_MORE) {
+    pj.error_code = stage2_is_ok;
+    if (reallocated) {
+      aligned_free((void *)buf);
+    }
+    return pj.error_code;
+  }
+  bool need_stage1 = false;
+  while (idx + window < len) {
+    actual_window = window;
+    if(need_stage1) { // first time around, we don't need stage 1
+      stage1_is_ok = find_structural_bits<T>(buf + idx, actual_window, pj, true);
+      if (stage1_is_ok != simdjson::SUCCESS) {
+        pj.error_code = stage1_is_ok;
+        if (reallocated) {
+          aligned_free((void *)buf);
+        }
+        return pj.error_code;
+      }
+    } else {
+      need_stage1 = true; // next round, we are going to need stage 1
+    }
+    last_index_t last_open_brace_index =
+        find_last_open_brace(buf + idx, actual_window, pj);
+    if (last_open_brace_index.has_error) {
+      pj.error_code = simdjson::WINDOW_ERROR;
+      if (reallocated) {
+        aligned_free((void *)buf);
+      }
+      return pj.error_code;
+    }
+    actual_window = last_open_brace_index.last_buf_index;
+    stage2_is_ok = unified_machine_continue<T>(
+        buf + idx, actual_window, pj, last_open_brace_index.last_index);
+    if (stage2_is_ok != simdjson::SUCCESS_AND_HAS_MORE) {
+      pj.error_code = stage2_is_ok;
+      if (reallocated) {
+        aligned_free((void *)buf);
+      }
+      return stage2_is_ok;
+    }
+    idx += actual_window;
+  }
+  if (idx < len) {
+    actual_window = len - idx;
+    if(need_stage1) { // we could make it at this point without needing stage 1
+      stage1_is_ok = find_structural_bits<T>(buf + idx, actual_window, pj, false);
+      if (stage1_is_ok != simdjson::SUCCESS) {
+        pj.error_code = stage1_is_ok;
+        if (reallocated) {
+          aligned_free((void *)buf);
+        }
+        return pj.error_code;
+      }
+    }
+    stage2_is_ok =
+        unified_machine_finish<T>(buf + idx, actual_window, pj);
+    idx += actual_window; // should be unnecessary
+  }
+  pj.error_code = stage2_is_ok;
+  if (reallocated) {
+    aligned_free((void *)buf);
+  }
+  return pj.error_code;
+}
+
+// like json_parse but parses the document in chunks of "window" bytes
+int interleaved_json_parse(const uint8_t *buf, size_t len, ParsedJson &pj,
+               size_t window, bool realloc_if_needed = true);
+int interleaved_json_parse(const char *buf, size_t len, ParsedJson &pj,
+               size_t window, bool realloc_if_needed = true);
+inline int interleaved_json_parse(const std::string &s, ParsedJson &pj, size_t window) {
+  return interleaved_json_parse(s.data(), s.length(), pj, window, true);
+}
+inline int interleaved_json_parse(const padded_string &s, ParsedJson &pj, size_t window) {
+  return interleaved_json_parse(s.data(), s.length(), pj, window, false);
+}
 } // namespace simdjson
 #endif