--- a/include/simdjson/portability.h
+++ b/include/simdjson/portability.h
@@ -9,14 +9,12 @@
 #include <iso646.h>
 #endif
 
-
 #if defined(__x86_64__) || defined(_M_AMD64)
 #define IS_X86_64 1
 #endif
 #if defined(__aarch64__) || defined(_M_ARM64)
 #define IS_ARM64 1
 #endif
-
 
 // this is almost standard?
 #undef STRINGIFY_IMPLEMENTATION_