--- a/src/arm64/stage2_build_tape.h
+++ b/src/arm64/stage2_build_tape.h
@@ -21,7 +21,12 @@ namespace simdjson {
 template <>
 WARN_UNUSED int
 unified_machine<Architecture::ARM64>(const uint8_t *buf, size_t len, ParsedJson &pj) {
-  return arm64::unified_machine(buf, len, pj);
+  arm64::stage2_status status;
+  int ret = arm64::unified_machine_init(buf, len, pj, status);
+  while(ret == simdjson::SUCCESS_AND_HAS_MORE) {
+    ret = arm64::unified_machine_continue(buf, len, pj, status, pj.n_structural_indexes);
+  }
+  return ret;
 }
 
 template <>