Currently stage 2 is mostly a bulky function. Moreover, it can only process a whole JSON document (or nothing).

In PR https://github.com/lemire/simdjson/pull/406, it was rewritten using a functions and an object-oriented approach, but with a significant performance penalty.

This new PR tries to simplify stage 2 and add the ability to process only part of a JSON document with (theoretically) the ability to resume parsing the rest of the document.

There are many applications of such an approach. One is parallelization as in PR https://github.com/lemire/simdjson/pull/406, another one is to do away with the need for whole string padding.

The idea here is that you can stop parsing before the beginning of an array or an object. So there are only two states to keep track off... either you are at the start of the array or at the start of an object.

This PR has a small performance penalty. However, it may be a penalty that we are willing to pay for the improved flexibility.

This PR is deliberately minimalist: instead of doing massive changes, I try to change as little as possible.

Comments are invited.

cc @jkeiser @piotte13 @pauldreik