@lemire 2019-12-24
The performance gap with the existing code is now a fraction of a cycle per structural character.