@jkeiser 2019-12-24
It doesn't seem simpler by itself, am I missing something? The algorithm is intentionally unchanged, which is fine, but It stores the same data internally (i/idx/c).

I also don't understand the use case, at least based on the code. I had thought it was so you could do the work in chunks. But using array/object begin as a chunk delimiter doesn't seem to give you much control over the chunk size--it depends entirely on your document. Most of the "pause/resume" use cases I can think of are trying to break up work as a sort of cooperative multitasking or to limit work size to a cache.

Why the focus on array/hash? Is it that we haven't found a performant way to to allow pause/resume at any point, or perhaps at "I expect to encounter a value now" points like after start of file, `{` `[` `,` and `:`?