@lemire 2019-12-25
The change is small because I want to progress without throwing away the performance. You can rewrite the code with a new design and then hope that the performance of the two final code paths is going to be similar, but you might be disappointed. Re-optimizing a new code path after you lost performance is hard. I prefer to work incrementally and make sure I understand the performance consequences of all changes.

Currently, the new code in this PR is still of lower performance. It shouldn't be. Even my small changes had a cost. I am working on narrowing it down.

What is the overarching design?

You have two functions. You have an initialization function and then you have a "continue" function which can be called as many times as you'd like. So you do stage 1 over a small block, identify a subwindow within the block that can be safely sent to stage 2, then run stage 2 initialization and stage 2 "continue" once, then you do another stage 1 starting from the end of the subwindows over another block, figure out a new subwindow within the new block,  and then again stage 2 "continue". Rinse and repeat.

Why the subwindows? Why not send predefined blocks of say 4kB to stage 1 and then stage 2? Because that's not safe. Stage 1 should not start or end in the middle of a string. So, at a minimum, after running stage 1, you must backtrack up to the last string detected... you cannot start a new stage 1 just anywhere. And if your stage 1 is configured to run over 4kB blocks and you have 1GB strings, you have a problem. There is no way around that, but we know that in practice most JSON documents have lots of objects/arrays at regular intervals. 

Note that stage 2 must stop processing at the designated structural character. This means that, during stage 2, you might need to check whether each and every structural character is the last one. That's an overhead. Remember that we use fewer than 10 cycles per structural character... far fewer sometimes... so an overhead per structural character is a serious consideration.

Furthermore, the code must be structured so that you can resume processing at any of the allowed points, which puts a constraint on how you can code things up.

Jérémie's code is able to suspend/resume at any structural character (so ',', '"', '[' and so forth), but in the new design I am currently planning to suspend/resume only at the start of array and objects. My design has lower overhead. Being able to stop/resume anywhere is more costly. It is less costly because only the start of arrays and objects have to pay an overhead. It is also less costly because our stage 2 code is already designed around the start of arrays and objects... so there is no need to rewrite the stage 2 code.

It is a trade-off. Everything single suspend/resume point comes with a performance. If you can suspend result at *any* point, then there is a greater overhead (at least on paper... miraculous implementations are always possible if one is sufficiently clever).

There is no way to define the window size and process strictly in terms of predefined window size. However, you can interleave stage 1 and stage 2. Run stage 1 over window A, concurrently run stage 2 over window A and stage 1 over the next block (B) , concurrently run stage 2 over window B and stage 1 over the next block (C) and so forth. In an ideal world where stage 1 and stage 2 both represent exactly 50% of the running cost, you'd go twice as fast. In practice, we expect more something of the order of a 50% boots based on Jérémie's prototype.

But you don't even need stage 1/stage 2 parallelization to see benefits. We know that for large files, we get hit badly by page walking and memory bandwith. So processing stage 1 over a page while prefetching the next page, processing stage 2, processing stage 1 over the newly loaded page, and so forth... should be much faster already for files significantly larger than 8kB.

Note that UTF-8 validation, in the future, should run as a separate step over the string buffer... and could be parallelized as well. So you could hit the large JSON with three cores... one doing stage 1, one doing stage 2 and one doing UTF-8 validation.


Anyhow, for now I just want an interruptible stage 2 that comes "nearly for free", ideally "entirely for free". This is the building block we need.