@lemire 2019-12-26
> It doesn't seem simpler by itself, am I missing something? 

Simpler is maybe not the right adjective. Less monolithic would be a better description of the intention.