@lemire 2019-12-26
The new code has a penalty of more or less on instruction per structural character... whenever the input has strings. It seems that something having to do with how strings are handled is slower with the new code. (There are thousands of assembly instructions affected by this change, so it is hard to tell where the issue is just by looking at the assembly.)