@jkeiser 2019-12-26
> The new code has a penalty of more or less on instruction per structural character... whenever the input has strings.

Does it make a difference if the strings are object keys (say `{ "abc": 123 }`, or does it happen for any string anywhere?