@lemire 2019-12-27
> Does it make a difference if the strings are object keys (say { "abc": 123 }, or does it happen for any string anywhere?

I don't know at this point. I didn't expect any difference. I am puzzled by the performance degradation.