@lemire 2019-12-27
> Does it make a difference if the strings are object keys (say { "abc": 123 },

Yes. This scenario alone is bad for us.