@lemire 2019-12-27
So with a document as silly as this...

```
[
{ "abc": null },
{ "abc": null },
{ "abc": null },
{ "abc": null },
{ "abc": null },
{ "abc": null },
(...)
```

I get a penalty of about 1 instruction per structural element. I don't know where the hurt is happening.