@lemire 2019-12-27
Same with this...

```
[
 "abc",
 "abc",
 "abc",
 "abc",
(...)
```