@lemire 2019-12-27
In fact, in the large array of strings, I get 1.5 extra instructions per strutural character. This means that I get maybe 3 extra instructions per string.