@lemire 2019-12-27
It is not a branch misprediction story. It is a flat out "using more instructions" story. I don't understand where the extra instructions come from. They seem to only show up with strings.