@lemire 2019-12-27
Ah. I managed to reproduce the same effect without using strings.