@lemire 2019-12-27
Ah. Might not be related to strings after all. 

The following is fine...

```
[
 "1",
 "1",
 "1",
 "1",
 "1",
 "1",
 "1",
 "1",
 "1",
```