@lemire 2019-12-27
Performance is now pretty much on par with master.