@jkeiser 2019-12-27
I'll give it a run locally.

I think I understand why you cut it where you did: to avoid a branch, you need to be able to choose a single entry point into the maze of gotos, so you basically chose about_to_begin_object_or_array. How come you didn't choose something like scope_continue? That would let you cut it off after any `,`, `[` or `{`.