@lemire 2019-12-31
I have ported the key elements of @piotte13's threaded approach into this PR. The result is ugly and fragile code but it is fast. The code is ugly, but there is not a lot of it, it is also not very complicated, it is just ugly because it needs love.

While Jérémie only got performance benefits for large files, I am getting speed boosts for small files like twitter... I reach nearly 2.8 GB/s (up from 2.4 GB/s for the non-threaded version).

new approach (with 130000 blocks)
```
$ ./parse -I 130000  jsonexamples/twitter.json
All Stages
|    Speed        :  23.1713 ns per block ( 97.47%) -   0.3621 ns per byte -   4.1376 ns per structural -    2.762 GB/s
```

old approach
```
$ ./parse  jsonexamples/twitter.json
All Stages
|    Speed        :  26.3292 ns per block ( 98.56%) -   0.4114 ns per byte -   4.7015 ns per structural -    2.431 GB/s
```


I can also reach really high speeds with some files... nearly 5 GB/s with gsoc...

```
$ ./parse -I 200000   jsonexamples/gsoc-2018.json
|    Speed        :  13.2363 ns per block ( 96.64%) -   0.2068 ns per byte -   9.0750 ns per structural -    4.835 GB/s
```

This is up from a respectable 3.6 GB/s.
```
./parse    jsonexamples/gsoc-2018.json

All Stages
|    Speed        :  17.7361 ns per block ( 98.33%) -   0.2771 ns per byte -  12.1602 ns per structural -    3.608 GB/s
```

Currently, this approach is tricky because of UTF-8 validation... whenever stage 1 ends in the middle of UTF8 code point, it fails. This is an issue that was already present in the code before I started doing this work. I am not worried about this, since we are going to fix this by moving UTF-8 validation out of the way anyhow. (It could be disabled for a time.)

Another tricky factor is that it is hard to pick the right window size. If you pick wrong, you get performance that is subpar compared to the single threaded version. That's expected in a way...

Furthermore, some preset window size will just lead to a failure when it can't find a '{' or '[' character in a window. Also, if you have the wrong type of file where the characters  '{' or '[' are hard to find, you will just waste time searching for them.

Anyhow, this stands as a proof of concept that we can substantially  boost the performance even on modest files.